+++ webpack.config.js
@@ -3,16 +3,6 @@ const webpack = require("webpack");
 const { BundleAnalyzerPlugin } = require("webpack-bundle-analyzer");
 const { merge } = require("webpack-merge");
 
-// TODO: Remove with #2273
-const bnJsReplaces = [
-  "tiny-secp256k1",
-  "asn1.js",
-  "create-ecdh",
-  "miller-rabin",
-  "public-encrypt",
-  "elliptic",
-];
-
 function getDefaultConfiguration() {
   return {
     cache: true,
@@ -26,16 +16,6 @@ function getDefaultConfiguration() {
     plugins: [
       new webpack.ProvidePlugin({ process: "process/browser" }),
       new webpack.ProvidePlugin({ Buffer: ["buffer", "Buffer"] }),
-      // TODO: Remove with #2273
-      new webpack.NormalModuleReplacementPlugin(/^bn.js$/, (resource) => {
-        if (
-          bnJsReplaces.some((pkg) =>
-            resource.context.includes(`node_modules/${pkg}`)
-          )
-        ) {
-          resource.request = "diffie-hellman/node_modules/bn.js";
-        }
-      }),
     ],
     module: {
       rules: [
@@ -53,8 +33,6 @@ function getDefaultConfiguration() {
       symlinks: false,
       fallback: {
         buffer: require.resolve("buffer"),
-        stream: require.resolve("stream-browserify"),
-        crypto: require.resolve("crypto-browserify"),
       },
     },
   };